feat: Extract LanguageSection module from ConfigScreen (v1.0.54)
Phase 4.8 Complete: Largest and most complex extraction yet!

Created src/config/ui/LanguageSection.js (512 lines) to handle:
- Dual-column language selector (enabled/available)
- Full drag-and-drop state machine with touch/mouse support
- Tap-to-toggle between columns
- Drag-to-reorder within columns
- Drag-to-move between columns
- Visual drop indicators and column highlighting
- Sophisticated drag threshold detection (10px movement)
- Request animation frame for smooth dragging
- Priority display (#1, #2, etc. for enabled languages)

Changes:
- NEW: src/config/ui/LanguageSection.js (512 lines - LARGEST MODULE!)
  - render() generates language selection HTML
  - populateLanguageColumns() dynamically builds language items
  - createLanguageElement() builds draggable language cards with flags
  - setupLanguageDragAndDrop() initializes drag system
  - makeLanguageItemDraggable() implements complex drag state machine
  - toggleLanguageColumn() handles tap-to-toggle
  - findDropTarget() calculates drop position from mouse/touch coords
  - updateDropIndicator() shows visual feedback during drag
  - clearDropIndicators() cleans up indicators
  - handleLanguageDrop() routes to reorder or move
  - reorderLanguage() handles same-column reordering
  - moveLanguage() handles cross-column movement

- UPDATED: src/config/ConfigScreen.js (1,329 → 871 lines, 35% reduction!)
  - Import LanguageSection module
  - Initialize this.languageSection in createUI() before HTML generation
  - Replace createLanguageSection() with this.languageSection.render()
  - Update populateLanguageColumns() calls to use languageSection instance
  - Removed 12 language methods (458 lines total):
    * createLanguageSection()
    * populateLanguageColumns()
    * createLanguageElement()
    * setupLanguageDragAndDrop()
    * makeLanguageItemDraggable()
    * toggleLanguageColumn()
    * findDropTarget()
    * updateDropIndicator()
    * clearDropIndicators()
    * handleLanguageDrop()
    * reorderLanguage()
    * moveLanguage()

- UPDATED: package.json
  - Version bumped to 1.0.54

Impact:
- Extracted 458 lines to separate module (35% reduction this phase!)
- LanguageSection is the largest, most complex module (512 lines)
- Sophisticated drag-and-drop UX preserved perfectly
- All 902 tests passing ✅

Progress:
- ConfigScreen.js: 3,284 → 871 lines (73% total reduction!!!)
- ALL 7 UI sections extracted to modules
- Next: Phase 4.9 will refactor ConfigScreen into a pure orchestrator (~220 lines target)

🎉 This is a major milestone - all content sections extracted!
